refactor(streams): move log-stream core behind EventSink
