+++ apps/desktop/src-tauri/src/lib.rs
@@ -21,8 +21,9 @@ use exec::{exec_close, exec_input, exec_resize, start_pod_exec, ExecManager};
 use files::{pick_kubeconfig_files, save_pasted_kubeconfig, save_text_file};
 use forward::{start_port_forward, stop_port_forward, ForwardManager};
 use helm::{helm_op_close, start_helm_op};
+use logs::{start_log_stream, stop_log_stream};
 use srelens_streams::helm::HelmManager;
-use logs::{start_log_stream, stop_log_stream, LogStreamManager};
+use srelens_streams::logs::LogStreamManager;
 use mcp::{
     install_srelens_cli, mcp_http_start, mcp_http_status, mcp_http_stop, srelens_cli_status,
     McpHttpManager,
