+++ crates/streams/src/lib.rs
@@ -3,6 +3,7 @@
 //! [`EventSink`] implemented by the host (Tauri events, WebSocket frames).
 
 pub mod helm;
+pub mod logs;
 pub mod sink;
 pub mod test_util;
